spatula sources into core
